Move rest of the forms to forms folder
Move SelectMultipleFilterForm and SelectMultipleFilterPlainForm to the
forms folder.
